final_no_bonus

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: aldgonza <aldgonza@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/10 20:09:32 by aldgonza          #+#    #+#             */
-/*   Updated: 2022/10/17 17:02:21 by aldgonza         ###   ########.fr       */
+/*   Updated: 2022/10/17 17:12:19 by aldgonza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@
 
 char	*get_next_line(int fd);
 int		ft_strlen(const char *str);
-size_t	ft_strlcpy(char *dest, char *src, size_t size);
+int		ft_strlcpy(char *dest, char *src, size_t size);
 char	*ft_strjoin(char *s1, char *s2);
 int		ft_strchr(const char *str, int ch);
 char	*ft_add_buff(char *buffer, int fd);
